pre-versions/v2/requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.26.0
pytest>=7.0.0
reportlab>=4.0.0